+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+# [v1.0.0-beta.47](https://github.com/TomokiMiyauci/react-hookable/compare/v1.0.0-beta.46...v1.0.0-beta.47) (2021-10-22)
+
+## ✨ New Features
+- [`7ba7fae`](https://github.com/TomokiMiyauci/react-hookable/commit/7ba7fae)   Add onNonShortcut feature to &#x60;useShortcutEffect&#x60;
+
+# [v1.0.0-beta.46](https://github.com/TomokiMiyauci/react-hookable/compare/v1.0.0-beta.45...v1.0.0-beta.46) (2021-10-22)
+
+## 🐛 Bug Fixes
+- [`5f5ae00`](https://github.com/TomokiMiyauci/react-hookable/commit/5f5ae00)   Fix export format as named export
+
 # [v1.0.0-beta.45](https://github.com/TomokiMiyauci/react-hookable/compare/v1.0.0-beta.44...v1.0.0-beta.45) (2021-10-22)
 
 ## ✨ New Features